Separate holder class from adapter class
