+++ ninchatsdk/src/main/java/com/ninchat/sdk/adapters/NinchatQueueListAdapter.java
@@ -1,55 +1,23 @@
 package com.ninchat.sdk.adapters;
 
 import android.app.Activity;
-import android.os.Handler;
-import android.os.Looper;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-
-import com.ninchat.sdk.NinchatSessionManager;
 import com.ninchat.sdk.R;
 import com.ninchat.sdk.activities.NinchatQueueActivity;
+import com.ninchat.sdk.adapters.holders.NinchatQueueViewHolder;
 import com.ninchat.sdk.models.NinchatQueue;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
-public final class NinchatQueueListAdapter extends RecyclerView.Adapter<NinchatQueueListAdapter.NinchatQueueViewHolder> {
+public final class NinchatQueueListAdapter extends RecyclerView.Adapter<NinchatQueueViewHolder> {
 
-    final class NinchatQueueViewHolder extends RecyclerView.ViewHolder {
-
-        NinchatQueueViewHolder(final View itemView) {
-            super(itemView);
-        }
-
-        void bind(final NinchatQueue queue) {
-            final Button button = itemView.findViewById(R.id.queue_name);
-
-            // If queue is closed, disable button and set alpha for look & feel
-            if (queue.isClosed()) {
-                button.setEnabled(false);
-                button.setAlpha(0.5f);
-                button.setText(NinchatSessionManager.getInstance().getQueueName(queue.getName(), queue.isClosed()));
-            } else {
-                button.setAlpha(1f);
-                button.setText(NinchatSessionManager.getInstance().getQueueName(queue.getName()));
-                button.setOnClickListener(v -> {
-                    final Activity activity = activityWeakReference.get();
-                    if (activity != null) {
-                        activity.startActivityForResult(NinchatQueueActivity.getLaunchIntent(activity, queue.getId()),  NinchatQueueActivity.REQUEST_CODE);
-                    }
-                });
-            }
-        }
-    }
-
-    protected List<NinchatQueue> queues = new ArrayList<>();
-    protected WeakReference<Activity> activityWeakReference;
+    private List<NinchatQueue> queues = new ArrayList<>();
+    private WeakReference<Activity> activityWeakReference;
 
     public NinchatQueueListAdapter(final Activity activity, final List<NinchatQueue> queues) {
         this.activityWeakReference = new WeakReference<>(activity);
@@ -59,36 +27,37 @@ public final class NinchatQueueListAdapter extends RecyclerView.Adapter<NinchatQ
     @NonNull
     @Override
     public NinchatQueueViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        return new NinchatQueueViewHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.item_queue, parent, false));
+        return new NinchatQueueViewHolder(
+                LayoutInflater.from(parent.getContext()).inflate(R.layout.item_queue, parent, false));
     }
 
     @Override
     public void onBindViewHolder(@NonNull NinchatQueueViewHolder holder, int position) {
-        holder.bind(queues.get(position));
+        holder.bind(queues.get(position), callback);
     }
 
     @Override
     public int getItemCount() {
-        return queues.size();
+        return queues == null ? 0 : queues.size();
     }
 
-    public void clear() {
-        this.queues.clear();
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
-            @Override
-            public void run() {
-                notifyDataSetChanged();
-            }
-        });
+    public void clearData() {
+        queues.clear();
+        notifyDataSetChanged();
     }
 
-    public void addQueue(final NinchatQueue queue) {
-        this.queues.add(queue);
-        new Handler(Looper.getMainLooper()).post(new Runnable() {
-            @Override
-            public void run() {
-                notifyItemInserted(queues.size() - 1);
-            }
-        });
+    public void addData(final NinchatQueue queue) {
+        final int previousSize = queues.size();
+        queues.add(queue);
+        notifyItemRangeInserted(previousSize, queues.size());
     }
+
+    private final NinchatQueueViewHolder.Callback callback = queueId -> {
+        final Activity activity = activityWeakReference.get();
+        if (activity != null) {
+            activity.startActivityForResult(
+                    NinchatQueueActivity.getLaunchIntent(activity, queueId),
+                    NinchatQueueActivity.REQUEST_CODE);
+        }
+    };
 }
